# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Crewquarters Control API
  description: 'Control plane for Crewquarters. JSON is camelCase. Errors use `{error: {code, message,
    requestId, details}}`. State-changing requests need the session cookie, an allowed Origin, and the
    `X-CSRF-Token` header. Mutating requests accept an `Idempotency-Key` header.'
  version: 0.1.0
paths:
...
  /api/v1/catalog/agents:
    get:
      tags:
      - agents
      summary: List catalog agents (ordered by agent id)
      operationId: list_catalog_api_v1_catalog_agents_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_CatalogAgentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/catalog/agents/{agent_id}:
    get:
      tags:
      - agents
      summary: Get a catalog agent
      operationId: get_catalog_agent_api_v1_catalog_agents__agent_id__get
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAgentOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
  /api/v1/catalog/agents/{agent_id}/versions/{version}:
    get:
      tags:
      - agents
      summary: Get one immutable agent version
      operationId: get_catalog_version_api_v1_catalog_agents__agent_id__versions__version__get
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      - name: version
        in: path
        required: true
        schema:
          type: string
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVersionOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
  /api/v1/catalog/agents/import:
    post:
      tags:
      - agents
      summary: Import a locally built agent (crewctl publish --target local)
      operationId: import_agent_api_v1_catalog_agents_import_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogImportIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAgentOut'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/agent-installations:
    get:
      tags:
      - agents
      summary: List installed agents
      operationId: list_installations_api_v1_agent_installations_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_InstallationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - agents
      summary: Install an agent, approving its exact permissions
      operationId: create_installation_api_v1_agent_installations_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationCreateIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
  /api/v1/agent-installations/{installation_id}:
    get:
      tags:
      - agents
      summary: Get an installed agent with readiness
      operationId: get_installation_api_v1_agent_installations__installation_id__get
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    patch:
      tags:
      - agents
      summary: Update configuration, enablement, version, or permission approval
      operationId: update_installation_api_v1_agent_installations__installation_id__patch
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationPatchIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    delete:
      tags:
      - agents
      summary: Uninstall an agent (soft delete; blocked while runs are active)
      operationId: delete_installation_api_v1_agent_installations__installation_id__delete
      parameters:
      - name: installation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Installation Id
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
components:
  schemas:
...
    AgentVersionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        version:
          type: string
          title: Version
        image:
          type: string
          title: Image
        imageDigest:
          type: string
          title: Imagedigest
        sdkProtocol:
          type: string
          title: Sdkprotocol
        architectures:
          items:
            type: string
          type: array
          title: Architectures
        triggers:
          items:
            type: string
          type: array
          title: Triggers
        permissions:
          additionalProperties: true
          type: object
          title: Permissions
        resources:
          additionalProperties: true
          type: object
          title: Resources
        configurationSchema:
          additionalProperties: true
          type: object
          title: Configurationschema
        resultSchema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Resultschema
        compatible:
          type: boolean
          title: Compatible
        compatibilityIssues:
          items:
            type: string
          type: array
          title: Compatibilityissues
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - version
      - image
      - imageDigest
      - sdkProtocol
      - architectures
      - triggers
      - permissions
      - resources
      - configurationSchema
      - resultSchema
      - compatible
      - compatibilityIssues
      - createdAt
      title: AgentVersionOut
...
    CatalogAgentOut:
      properties:
        agentId:
          type: string
          title: Agentid
        name:
          type: string
          title: Name
        summary:
          type: string
          title: Summary
        publisher:
          type: string
          title: Publisher
        source:
          type: string
          enum:
          - bundled
          - imported
          title: Source
        trustStatus:
          type: string
          enum:
          - curated
          - imported_unreviewed
          title: Truststatus
        currentVersion:
          type: string
          title: Currentversion
        versions:
          items:
            type: string
          type: array
          title: Versions
        latest:
          $ref: '#/components/schemas/AgentVersionOut'
        installed:
          type: boolean
          title: Installed
      type: object
      required:
      - agentId
      - name
      - summary
      - publisher
      - source
      - trustStatus
      - currentVersion
      - versions
      - latest
      - installed
      title: CatalogAgentOut
    CatalogImportIn:
      properties:
        manifest:
          additionalProperties: true
          type: object
          title: Manifest
          description: Agent manifest (crewquarters/v1alpha1).
      type: object
      required:
      - manifest
      title: CatalogImportIn
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InstallationCreateIn:
      properties:
        agentId:
          type: string
          title: Agentid
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
          description: Defaults to the catalog's current version.
        config:
          additionalProperties: true
          type: object
          title: Config
        approvedPermissions:
          additionalProperties: true
          type: object
          title: Approvedpermissions
          description: Must equal the version's requested permissions exactly.
        modelBindings:
          additionalProperties:
            type: string
          type: object
          title: Modelbindings
          description: Owner choice of variant per requested profile family.
          examples:
          - local.general: local.general.quality
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - agentId
      - approvedPermissions
      title: InstallationCreateIn
    InstallationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        agentId:
          type: string
          title: Agentid
        agentName:
          type: string
          title: Agentname
        agentVersion:
          type: string
          title: Agentversion
        agentVersionId:
          type: string
          format: uuid
          title: Agentversionid
        config:
          additionalProperties: true
          type: object
          title: Config
        requestedPermissions:
          additionalProperties: true
          type: object
          title: Requestedpermissions
        approvedPermissions:
          additionalProperties: true
          type: object
          title: Approvedpermissions
        capabilities:
          items:
            type: string
          type: array
          title: Capabilities
        modelBindings:
          additionalProperties:
            type: string
          type: object
          title: Modelbindings
        needsReapproval:
          type: boolean
          title: Needsreapproval
        enabled:
          type: boolean
          title: Enabled
        version:
          type: integer
          title: Version
        readiness:
          $ref: '#/components/schemas/Readiness'
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - agentId
      - agentName
      - agentVersion
      - agentVersionId
      - config
      - requestedPermissions
      - approvedPermissions
      - capabilities
      - modelBindings
      - needsReapproval
      - enabled
      - version
      - readiness
      - createdAt
      - updatedAt
      title: InstallationOut
...
    Page_CatalogAgentOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CatalogAgentOut'
          type: array
          title: Items
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
          description: Opaque cursor for the next page; null on the last page.
      type: object
      required:
      - items
      title: Page[CatalogAgentOut]
...
    Page_InstallationOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/InstallationOut'
          type: array
          title: Items
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
          description: Opaque cursor for the next page; null on the last page.
      type: object
      required:
      - items
      title: Page[InstallationOut]
...
    Readiness:
      properties:
        ready:
          type: boolean
          title: Ready
        checks:
          items:
            $ref: '#/components/schemas/ReadinessCheck'
          type: array
          title: Checks
      type: object
      required:
      - ready
      - checks
      title: Readiness
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: cq_session
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
    service:
      type: http
      scheme: bearer
      description: Internal service credential
servers:
- url: /
  description: Same origin as the web UI